refactor: :recycle: Move logic Login Page to a component
