--- src/pages/Login/Login.css
@@ -1,59 +1,0 @@
-.login-container {
-  width: 100%;
-  display: flex;
-  margin-top: -150px;
-  align-items: center;
-  justify-content: center;
-}
-
-.login-card {
-  background-color: #f1e4ee;
-  width: 360px;
-  height: 410px;
-  padding: 20px;
-  border-radius: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 20px;
-}
-
-.login-form {
-  width: 80%;
-}
-
-.login-card > span {
-  font-weight: bold;
-  font-size: 17px;
-}
-
-.login-card > a {
-  color: #049dbf;
-  font-weight: 600;
-}
-
-.login-button {
-  background-color: #045f98;
-  color: white;
-  height: 40px;
-  border-radius: 5px;
-  width: 80%;
-  display: flex;
-  align-items: center;
-}
-
-.login-button > img {
-  width: 7%;
-  height: auto;
-}
-
-.login-button > span {
-  width: 100%;
-  margin-right: 7%;
-  font-size: 15px;
-  font-weight: 100;
-}
-.sidebar {
-  display: none;
-}
